test: split browser clipper route coverage
